move dashboard info to separate fragment
