+++ sampleapp/src/main/res/values/strings.xml
@@ -1,5 +1,5 @@
 <resources>
-<!--    <string name="app_name">SampleApp</string>   See build.gradle buildTypes -->
+    <!--    <string name="app_name">SampleApp</string>   See build.gradle buildTypes -->
     <string name="navigation_label">TOPL-Android Demo</string>
 
     <!-- HomeFragment -->
@@ -9,4 +9,6 @@
     <string name="button_new_identity">New Identity</string>
     <string name="button_debugging_disable">Disable Debugging</string>
     <string name="button_debugging_enable">Enable Debugging</string>
+    <!-- TODO: Remove or change this placeholder text -->
+    <string name="hello_blank_fragment">Hello blank fragment</string>
 </resources>
